backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.12.3
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.27.0
exceptiongroup==1.3.0
flake8==7.3.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
msgpack==1.0.7
Nuitka==2.7.12
numpy==1.26.3
ordered-set==4.1.0
packaging==25.0
pillow==10.2.0
psutil==7.0.0
pluggy==1.6.0
maturin==1.9.2
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.26.0
soupsieve==2.7
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
watchdog==3.0.0
zstandard==0.23.0